+++ app/vmui/packages/vmui/src/components/Main/Modal/Modal.tsx
@@ -67,11 +67,11 @@ const Modal: FC<ModalProps> = ({
       })}
       onMouseDown={onClose}
     >
-      <div className="vm-modal-content">
-        <div
-          className="vm-modal-content-header"
-          onMouseDown={handleMouseDown}
-        >
+      <div
+        className="vm-modal-content"
+        onMouseDown={handleMouseDown}
+      >
+        <div className="vm-modal-content-header">
           {title && (
             <div className="vm-modal-content-header__title">
               {title}
@@ -91,7 +91,6 @@ const Modal: FC<ModalProps> = ({
         {/* tabIndex to fix Ctrl-A */}
         <div
           className="vm-modal-content-body"
-          onMouseDown={handleMouseDown}
           tabIndex={0}
         >
           {children}